+++ tests/evergreen/test_api.py
@@ -644,13 +644,21 @@ class TestTaskApi(object):
         mocked_api.task_by_id("task_id")
         expected_url = mocked_api._create_url("/tasks/task_id")
         mocked_api.session.request.assert_called_with(
-            url=expected_url, params=None, timeout=None, data=None, method="GET"
+            url=expected_url, params={}, timeout=None, data=None, method="GET"
         )
 
     def test_task_by_id_with_fetch_executions(self, mocked_api):
         mocked_api.task_by_id("task_id", fetch_all_executions=True)
         expected_url = mocked_api._create_url("/tasks/task_id")
         expected_params = {"fetch_all_executions": True}
+        mocked_api.session.request.assert_called_with(
+            url=expected_url, params=expected_params, timeout=None, data=None, method="GET"
+        )
+
+    def test_task_by_id_with_specific_execution(self, mocked_api):
+        mocked_api.task_by_id("task_id", execution=0)
+        expected_url = mocked_api._create_url("/tasks/task_id")
+        expected_params = {"execution": 0}
         mocked_api.session.request.assert_called_with(
             url=expected_url, params=expected_params, timeout=None, data=None, method="GET"
         )